+++ src/views/microdeposits/AccountInfo.js
@@ -5,7 +5,7 @@ import _isEmpty from 'lodash/isEmpty'
 import { useTokens } from '@kyper/tokenprovider'
 import { Text } from '@kyper/text'
 import { ChevronRight } from '@kyper/icon/ChevronRight'
-import { TextInput, SelectionBox } from 'src/privacy/input'
+import { TextField, SelectionBox } from 'src/privacy/input'
 import { Button } from '@mui/material'
 
 import useAnalyticsPath from 'src/hooks/useAnalyticsPath'
@@ -26,21 +26,6 @@ import {
 import { useForm } from 'src/hooks/useForm'
 import { getDelay } from 'src/utilities/getDelay'
 
-const schema = {
-  accountNumber: {
-    label: __('Account number'),
-    required: true,
-    pattern: 'digits',
-    equalTo: 'accountNumberConfirm',
-  },
-  accountNumberConfirm: {
-    label: __('Confirm account number'),
-    required: true,
-    pattern: 'digits',
-    equalTo: 'accountNumber',
-  },
-}
-
 export const AccountInfo = (props) => {
   const { accountDetails, focus, onContinue } = props
   const containerRef = useRef(null)
@@ -52,6 +37,20 @@ export const AccountInfo = (props) => {
   const initialForm = {
     accountNumber: accountDetails?.account_number ?? '',
     accountNumberConfirm: accountDetails?.account_number ?? '',
+  }
+  const schema = {
+    accountNumber: {
+      label: __('Account number'),
+      required: true,
+      pattern: 'digits',
+      equalTo: 'accountNumberConfirm',
+    },
+    accountNumberConfirm: {
+      label: __('Confirm account number'),
+      required: true,
+      pattern: 'digits',
+      equalTo: 'accountNumber',
+    },
   }
   const { handleTextInputChange, handleSubmit, values, errors } = useForm(
     handleContinue,
@@ -122,11 +121,13 @@ export const AccountInfo = (props) => {
 
         <SlideDown delay={getNextDelay()}>
           <div style={styles.inputStyle}>
-            <TextInput
+            <TextField
               autoComplete="off"
               autoFocus={focus === AccountFields.ACCOUNT_NUMBER}
-              data-test="account-number-input"
-              errorText={errors.accountNumber}
+              error={!!errors.accountNumber}
+              fullWidth={true}
+              helperText={errors.accountNumber}
+              inputProps={{ 'data-test': 'account-number-input' }}
               label={schema.accountNumber.label}
               name="accountNumber"
               onChange={handleTextInputChange}
@@ -136,10 +137,12 @@ export const AccountInfo = (props) => {
             />
           </div>
           <div style={styles.inputStyle}>
-            <TextInput
+            <TextField
               autoComplete="off"
-              data-test="confirm-account-number-input"
-              errorText={errors.accountNumberConfirm}
+              error={!!errors.accountNumberConfirm}
+              fullWidth={true}
+              helperText={errors.accountNumberConfirm}
+              inputProps={{ 'data-test': 'confirm-account-number-input' }}
               label={schema.accountNumberConfirm.label}
               name="accountNumberConfirm"
               onChange={handleTextInputChange}